engine: drafts verified on any composition -- verify segments, per-token state kept and accepted by position, ComposedModel speculative decode behind a Drafter protocol
base/composition: Segment.verify marks provisional tokens; a feature that carries per-sequence values records the
value after each of them (state.put(..., at=j), put_state) and State.accept(seq, n) keeps the first n -- rolling back
is choosing a position. forward(hidden=True) hands back the residual state before the closing mix (what Qwen3.8's MTP
reads). GDN/KDA (gated_delta_rule all_states, conv), the n-gram memory (context, dilated conv), PreNorm's output convs
and attention's k/v convs record per token; causal_conv.conv_states gives the conv state after each token.

base/composed: PositionStore keeps a ring of slot regions for a verify step and copies the accepted position's
values back; a checkpoint at a block boundary inside the step just accepted reads the ring (the runner checkpoints the
boundaries a decode step crossed). ComposedModel(drafter=...) verifies [last token] + drafts by position, sampling every
position with the uniform its generation count draws without drafts, so a drafter changes how many tokens a step
yields and never which; drafts are capped by the row's remaining budget; horizon = context + 1 + k.

tests/test_engine_speculative.py: accepting n == feeding n (reference State and PositionStore), the refusals, and runs
through the runner equal to runs without drafts token for token -- perfect, partly right and random drafts, greedy
and sampled, two rows, an end token inside an accepted run, min_tokens, a second turn, an adopted prefix, a boundary
checkpointed from the ring. tests/test_engine_selfchecks.py restores the default dtype the norm/linear/logits
self-checks set to bf16 (it leaked into every test run after them); qwen38 random_weights draws in fp32 explicitly.

+++ engine/profiles/qwen38/weights.py
@@ -184,7 +184,7 @@ def random_weights(cfg: dict, seed: int = 0, prefix: str = "model.") -> dict:
     (`experts.gate_up_proj` [E, 2I, H], `experts.down_proj` [E, H, I]) and the whole PLE table as one tensor."""
     from engine.modules.ngram_embedding import head_tables
     g = torch.Generator().manual_seed(seed)
-    r = lambda *shape, scale=1.0: torch.randn(*shape, generator=g) * scale
+    r = lambda *shape, scale=1.0: torch.randn(*shape, generator=g, dtype=torch.float32) * scale   # not the global default
     H, hc = cfg["hidden_size"], cfg["hc_count"]
     kd, vd, nk, nv = cfg["linear_key_head_dim"], cfg["linear_value_head_dim"], cfg["linear_num_key_heads"], cfg["linear_num_value_heads"]
     conv_dim = 2 * nk * kd + nv * vd
@@ -211,8 +211,8 @@ def random_weights(cfg: dict, seed: int = 0, prefix: str = "model.") -> dict:
             w[f"{la}.out_proj.weight"] = r(H, nv * vd, scale=0.1)
             w[f"{la}.norm.weight"] = 1 + r(vd, scale=0.1)
             w[f"{la}.conv1d.weight"] = r(conv_dim, 1, cfg["linear_conv_kernel_dim"], scale=0.3)
-            w[f"{la}.dt_bias"] = torch.ones(nv)
-            w[f"{la}.A_log"] = torch.empty(nv).uniform_(0.01, 16, generator=g).log()
+            w[f"{la}.dt_bias"] = torch.ones(nv, dtype=torch.float32)
+            w[f"{la}.A_log"] = torch.empty(nv, dtype=torch.float32).uniform_(0.01, 16, generator=g).log()
         else:
             sa = f"{base}.self_attn"
             w[f"{sa}.q_proj.weight"] = r(heads * hd * 2, H, scale=0.1)
